+++ package.json
@@ -46,6 +46,7 @@
     "react-helmet": "^6.1.0"
   },
   "devDependencies": {
+    "@testing-library/react": "^11.2.3",
     "babel-jest": "^26.6.3",
     "babel-preset-gatsby": "^0.10.0",
     "eslint": "^7.18.0",
@@ -57,8 +58,7 @@
     "eslint-plugin-react": "^7.22.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
-    "prettier": "^2.2.1",
-    "react-test-renderer": "^17.0.1"
+    "prettier": "^2.2.1"
   },
   "engines": {
     "node": ">=12"
