+++ src/components/resume/Contacts.spec.jsx
@@ -1,18 +1,16 @@
 import React from "react";
-import renderer from "react-test-renderer";
+import { render } from "@testing-library/react";
 
 import Contacts from "./Contacts.jsx";
 
 describe("Contacts Component", () => {
-  let wrapper;
-
   it("should render properly", () => {
-    wrapper = renderer.create(<Contacts />).toJSON();
-    expect(wrapper).toMatchSnapshot();
+    const { asFragment } = render(<Contacts />);
+    expect(asFragment()).toMatchSnapshot();
   });
 
   it("should render properly the alt version", () => {
-    wrapper = renderer.create(<Contacts isAltVersion />).toJSON();
-    expect(wrapper).toMatchSnapshot();
+    const { asFragment } = render(<Contacts isAltVersion />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });
