+++ src/components/resume/Education.spec.jsx
@@ -1,16 +1,11 @@
 import React from "react";
-import renderer from "react-test-renderer";
+import { render } from "@testing-library/react";
 
 import Education from "./Education.jsx";
 
 describe("Education Component", () => {
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = renderer.create(<Education />).toJSON();
-  });
-
   it("should render properly", () => {
-    expect(wrapper).toMatchSnapshot();
+    const { asFragment } = render(<Education />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });
