+++ src/components/resume/Extras.spec.jsx
@@ -1,16 +1,11 @@
 import React from "react";
-import renderer from "react-test-renderer";
+import { render } from "@testing-library/react";
 
 import Extras from "./Extras.jsx";
 
 describe("Extras Component", () => {
-  let wrapper;
-
-  beforeEach(() => {
-    wrapper = renderer.create(<Extras />).toJSON();
-  });
-
   it("should render properly", () => {
-    expect(wrapper).toMatchSnapshot();
+    const { asFragment } = render(<Extras />);
+    expect(asFragment()).toMatchSnapshot();
   });
 });
